fixed bugs for file separation
